requirements.txt
Flask
Pillow
numpy 
scikit-image
tensorflow
Gunicorn